refactor skill services
