+++ app/src/main/java/edu/kmaooad/capstone23/competences/services/impl/SkillSetServiceImpl.java
@@ -1,10 +1,10 @@
-package edu.kmaooad.capstone23.competences.services;
+package edu.kmaooad.capstone23.competences.services.impl;
 
 import edu.kmaooad.capstone23.competences.dal.SkillSet;
 import edu.kmaooad.capstone23.competences.dal.SkillSetRepository;
+import edu.kmaooad.capstone23.competences.services.SkillSetService;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
-import jakarta.enterprise.context.ApplicationScoped;
 import org.bson.types.ObjectId;
 
 import java.util.Optional;
